backend/requirements.txt
asgiref==3.2.10
django-filter
Django==3.0.5
djangorestframework-simplejwt==4.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4
python-dotenv
pytz==2020.1
requests
sqlparse==0.3.1
djoser